% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pushchart.R
\name{get_push_results}
\alias{get_push_results}
\title{Get push results for a set of point spreads}
\usage{
get_push_results(.pt_sprds, .scores_df, .wnd = 0)
}
\arguments{
\item{.pt_sprds}{numeric vector. Set of point spreads}

\item{.scores_df}{dataframe. Final scores, point spread, and season}

\item{.wnd}{window of points to include in the push results}
}
\value{
dataframe
}
\description{
Get push results for a set of point spreads
}
\examples{
/dontrun{
 get_push_results(1:14, nfl_scores, 1.5)
}
}
